requirements.txt
-c constraints.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==5.3.0
anchorpy==0.20.1
anchorpy-core==0.2.0
annotated-types==0.7.0
anyio==3.6.2
apischema==0.17.5
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==22.1.0
backoff==2.2.1
base58==2.1.1
based58==0.1.1
blinker==1.8.2
borsh-construct==0.1.0
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.2
construct==2.10.68
construct-typing==0.5.3
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.2.1
driftpy>=0.7.99
Events==0.5
exceptiongroup==1.0.4
executing==2.1.0
fastapi==0.115.0
fastapi-restful==0.6.0
fastapi-utils==0.7.0
fastjsonschema==2.20.0
flake8==6.0.0
fonttools==4.54.1
frozenlist==1.3.3
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
gunicorn==23.0.0
h11==0.14.0
httpcore>=1.0.0
httpx>=0.25.2
humanize==4.10.0
idna==3.4
iniconfig==1.1.1
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jito_searcher_client==0.1.4
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.17.3
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
loguru==0.6.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
more-itertools==8.14.0
multidict==6.0.3
mypy==1.10.0
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
OSlash==0.6.3
packaging==23.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.0.0
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==16.1.0
pycares==4.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pyflakes==3.0.1
Pygments==2.18.0
pyheck==0.1.5
pyinstrument==4.7.3
pyparsing==3.1.4
pyrsistent==0.19.2
pythclient==0.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==26.2.0
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
snakeviz==2.2.0
sniffio==1.3.0
solana==0.34.0
solders==0.21.0
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.6
streamlit==1.41.0
sumtypes==0.1a6
tenacity==8.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-cachetools==4.2.10
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.13
uvicorn==0.31.0
watchdog==4.0.1
wcwidth==0.2.13
websockets==10.4
wrapt==1.16.0
yarl==1.8.2
zstandard==0.18.0
